+++ lazy_propagation_set_max.cpp
@@ -43,28 +43,28 @@ template<typename T> using ordered_map = tree<T, int, less<T>, rb_tree_tag, tree
 //Lazy Propagation on Set Modifications, Max queries
 
 #define N 10000
-ll t[N << 2];
+int t[N << 2];
 bool mark[N << 2];
 
-void build(ll v, ll tl, ll tr, int arr[]){
+void build(int v, int tl, int tr, int arr[]){
     if(tl == tr){
         t[v] = arr[tl];
         return;
     }
-    ll tm = (tl+tr)>>1;
+    int tm = (tl+tr)>>1;
     build(v<<1,tl,tm,arr);
     build(v<<1|1,tm+1,tr,arr);
     t[v] = max(t[v<<1], t[v<<1|1]);
 }
 
-void push(ll v){
+void push(int v){
     t[v<<1] = t[v<<1|1] = t[v];
     mark[v<<1] = mark[v<<1|1] = 1;
     mark[v] = 0;
 
 }
 
-void modify(ll v, ll tl, ll tr, ll l, ll r, ll val){
+void modify(int v, int tl, int tr, int l, int r, int val){
     if(l > r) return;
     if(tl == l && tr == r){
         t[v] = val;
@@ -73,20 +73,20 @@ void modify(ll v, ll tl, ll tr, ll l, ll r, ll val){
     }
     if(mark[v])
         push(v);
-    ll tm = (tl+tr)>>1;
+    int tm = (tl+tr)>>1;
     modify(v<<1,tl,tm,l,min(tm,r),val);
     modify(v<<1|1,tm+1,tr,max(l,tm+1),r,val);
     t[v] = max(t[v<<1],t[v<<1|1]);
 }
 
-ll query(ll v, ll tl, ll tr, ll l, ll r){
+int query(int v, int tl, int tr, int l, int r){
     if(l > r) return -0x7f7f7f7f;
     if(tl == l && tr == r){
         return t[v];
     }
     if(mark[v])
         push(v);
-    ll tm = (tl+tr)>>1;
+    int tm = (tl+tr)>>1;
     return max(
     query(v<<1,tl,tm,l,min(tm,r)),
     query(v<<1|1,tm+1,tr,max(l,tm+1),r)
